Refactor - Move utility function to utils
We want to use them in Checker too.
